+++ experimental/packages/opentelemetry-instrumentation-grpc/src/grpc-js/index.ts
@@ -20,7 +20,7 @@ import {
   isWrapped,
 } from '@opentelemetry/instrumentation';
 import { InstrumentationBase } from '@opentelemetry/instrumentation';
-import { GrpcInstrumentationConfig } from '../types';
+import { GrpcInstrumentationConfig, metadataCaptureType } from '../types';
 import {
   ServerCallWithMeta,
   SendUnaryDataCallback,
@@ -49,17 +49,20 @@ import {
   getMetadata,
 } from './clientUtils';
 import { EventEmitter } from 'events';
-import { _extractMethodAndService } from '../utils';
+import {_extractMethodAndService, metadataCapture} from '../utils';
 import { AttributeValues } from '../enums/AttributeValues';
 import { SemanticAttributes } from '@opentelemetry/semantic-conventions';
 
 export class GrpcJsInstrumentation extends InstrumentationBase {
+  private _metadataCapture: metadataCaptureType;
+
   constructor(
     name: string,
     version: string,
     config?: GrpcInstrumentationConfig,
   ) {
     super(name, version, config);
+    this._metadataCapture = this._createMetadataCapture();
   }
 
   init() {
@@ -120,6 +123,11 @@ export class GrpcJsInstrumentation extends InstrumentationBase {
 
   override getConfig(): GrpcInstrumentationConfig {
     return super.getConfig();
+  }
+
+  override setConfig(config?: GrpcInstrumentationConfig): void {
+    super.setConfig(config);
+    this._metadataCapture = this._createMetadataCapture();
   }
 
   /**
@@ -298,8 +306,11 @@ export class GrpcJsInstrumentation extends InstrumentationBase {
             [SemanticAttributes.RPC_METHOD]: method,
             [SemanticAttributes.RPC_SERVICE]: service,
           });
+
+        instrumentation._metadataCapture.client.captureRequestMetadata(span, metadata);
+
         return context.with(trace.setSpan(context.active(), span), () =>
-          makeGrpcClientRemoteCall(original, args, metadata, this)(span)
+          makeGrpcClientRemoteCall(instrumentation._metadataCapture, original, args, metadata, this)(span)
         );
       }
       Object.assign(clientMethodTrace, original);
@@ -333,5 +344,16 @@ export class GrpcJsInstrumentation extends InstrumentationBase {
         );
       }
     });
+  }
+
+  private _createMetadataCapture(): metadataCaptureType {
+    const config = this.getConfig();
+
+    return {
+      client: {
+        captureRequestMetadata: metadataCapture('request', config.metadataToSpanAttributes?.client?.requestMetadata ?? []),
+        captureResponseMetadata: metadataCapture('response', config.metadataToSpanAttributes?.client?.responseMetadata ?? [])
+      }
+    };
   }
 }